Wire-up Place Order Functionality with UI

+++ api-client/src/main/java/com/kshitijpatil/tazabazar/api/dto/OrderLine.kt
@@ -6,7 +6,7 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 data class OrderLine(
     @Json(name = "inventoryId")
-    val inventoryId: Long,
+    val inventoryId: Int,
     @Json(name = "quantity")
     val quantity: Int
 )
